Completed Part 1: Graph Class
I implemented a Graph class. I created an add_vertex and add_edge functions. To test my implementation, I instantiated an empty graph and called both functions on it to test them.  As a stretch goal, I added checks to the graph to ensure that edges to nonexistent vertices are rejected.